Separated expenes_manager to api and frontend apps
