+++ code/apps/api/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from apps.api.models import Category
+
+class CategorySerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Category
+        fields = ['id', 'category_name']
