+++ code/apps/api/views.py
@@ -1,0 +1,8 @@
+from rest_framework import viewsets
+from .serializers import CategorySerializer
+from apps.api.models import Category
+
+class CategoryViewSet(viewsets.ModelViewSet):
+    queryset = Category.objects.all()
+    serializer_class = CategorySerializer
+
