--- code/apps/expenses_manager/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ExpenseManagerConfig(AppConfig):
-    name = 'expenses_manager'
