+++ code/apps/frontend/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path
 from . import views
 
-
-appname = 'expenses_manager'
+appname = 'frontend'
 urlpatterns = [
     path("stats", views.stats , name='stats'),
 
@@ -19,5 +18,4 @@ urlpatterns = [
     path("expenses/category/create", views.create_category , name='create category'),
     path("expenses/category/<int:pk>/delete", views.DeleteCategory.as_view() , name='delete category'),
     path("expenses/category/<int:category_id/update", views.create_category , name='update category'),
-    
 ]
