+++ code/budget_stat/settings.py
@@ -34,13 +34,15 @@ ALLOWED_HOSTS = []
 
 INSTALLED_APPS = [
     "apps.authentication",
-    "apps.expenses_manager",
+    "apps.api",
+    "apps.frontend",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
+    "rest_framework",
 ]
 
 MIDDLEWARE = [
